refactor: refactor content description
