+++ ui/src/main/java/com/baghdad/ui/feature/review/component/PlaceHolderReviewerImage.kt
@@ -1,4 +1,4 @@
-package com.baghdad.ui.feature.component.islamicImage
+package com.baghdad.ui.feature.review.component
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.size
@@ -6,6 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.baghdad.design_system.R
 import com.baghdad.design_system.component.Icon
@@ -18,7 +19,7 @@ fun PlaceHolderReviewerImage() {
     ) {
         Icon(
             painter = painterResource(R.drawable.img_reviewer_default),
-            contentDescription = "Default Image",
+            contentDescription = stringResource(com.baghdad.ui.R.string.default_image),
             tint = Theme.color.hint,
             modifier = Modifier.size(28.dp)
         )
